Introduce HitsInternal interface.
Current implementations are HitsArrays and HitsArrays32 (the old version),
which differ in how many results they can hold. Factory methods instantiate
the appropriate implementation.

+++ engine/src/main/java/nl/inl/blacklab/search/results/HitsFromQuery.java
@@ -194,8 +194,8 @@ public class HitsFromQuery extends Hits {
             }
             try {
                 boolean readAllHits = number < 0;
-                int maxHitsToCount = searchSettings.maxHitsToCount();
-                int maxHitsToProcess = searchSettings.maxHitsToProcess();
+                long maxHitsToCount = searchSettings.maxHitsToCount();
+                long maxHitsToProcess = searchSettings.maxHitsToProcess();
                 while (readAllHits || hitsArrays.size() < number) {
 
                     // Abort if asked
